refactor tests and clippy stuff
